requirements.txt
# Core Flask and Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Data Processing
polars==0.19.3
pyarrow==13.0.0
numpy==1.24.3
scipy==1.11.2
pandas==2.0.3

# Statistical Analysis
statsmodels==0.14.0
scikit-learn==1.3.0

# Configuration and Validation
pydantic==2.3.0
pydantic-settings==2.0.3
ruamel.yaml==0.17.32

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0

# Logging and Monitoring
loguru==0.7.2
structlog==23.1.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
httpx==0.25.0

# Parallel Processing
joblib==1.3.2
multiprocessing-logging==0.3.4

# Development
python-dotenv==1.0.0
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Optional for future distributed processing
# ray==2.6.3
# dask==2023.9.1